Split AlternativeNames into two files
