+++ TMDbLib/Objects/General/AlternativeNames.cs
@@ -11,13 +11,4 @@ namespace TMDbLib.Objects.General
         [JsonProperty("results")]
         public List<AlternativeName> Results { get; set; }
     }
-
-    public class AlternativeName
-    {
-        [JsonProperty("name")]
-        public string Name { get; set; }
-
-        [JsonProperty("type")]
-        public string Type { get; set; }
-    }
 }
